requirements.txt
python-telegram-bot==20.7
deep-translator==1.11.4
python-dotenv==1.0.1
gTTS==2.5.4
SpeechRecognition==3.10.4
pydub==0.25.1
